requirements.txt
Django==1.11
django-bootstrap3==9.0.0
djangorestframework==3.6.2
Pygments==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
